feat: add create embed fusion api (#1657)
* feat: add embed create/get/delete api interface definition

* feat: add enable swagger env

* feat: complete embedlink logic

* fix: api usage datasheet id null exception

* feat: add createEmbed link param validator

* feat: add deleted embed link param validator

* feat: add enterprise guard

* Squashed commit of the following:

commit 04bae792de13b1f716147ba182d3e2bf7d878b15
Author: Aria <sujian@vikadata.com>
Date:   Thu Nov 10 20:31:07 2022 +0800

    feat: new env variable for AWSC (#1836)

commit 647ed43b38ab5dbbabec05b5031c7d85d05ab9e3
Author: Aria <sujian@vikadata.com>
Date:   Thu Nov 10 19:53:44 2022 +0800

    chore: support custom header (#1806)

commit 5f2d9da0664a19e204892bedc503fde4e676b5c3
Author: 王开朗 <wangkailang@users.noreply.github.com>
Date:   Thu Nov 10 19:05:56 2022 +0800

    fix: datasheet next eslint (#1834)



commit acb3070c13ebd000a3cf8d72b9e5fe8964d789f3
Author: 刘子敬 <liuzijing@vikadata.com>
Date:   Thu Nov 10 18:56:56 2022 +0800

    refactor: translate backend-service model ro and vo (#1832)

    * refactor: translate model ro

    * refactor: translate model vo

commit 7e0267b262c181b8520cf1d3b5506155d68ed6fb
Author: Kelly Chan <kelly@vikadata.com>
Date:   Thu Nov 10 18:10:20 2022 +0800

    fix: i18n export  StringKeysType (#1820)

    * fix: i18n export  StringKeysType

    * fix: ERROR_STR delete

commit ca3f17c75d50eedc9e3fe4c7bd0593ffc8376fe3
Author: Kelly Chan <kelly@vikadata.com>
Date:   Thu Nov 10 16:11:00 2022 +0800

    refactor: i18n.class OOP refactor, and clean useless generated code. (#1814)

    * feat: init i18n.class

    * refactor: separate core/i18n

    * refactor: i18n refactor after @huanghao

    * chore: class public and private

    * refactor: code lint

    * feat: i18n language pack loader (for dynamic load)

    * chore: i18n README

    * refactor: move generate stringskey.interface.ts to "/core/config"

    * refactor: delete strings.interface.ts

    * refactor: move `i18n` to `exports` directory

    * fix: exports/i18n require error

commit 89dc35f341667039d8b190e42894774e503ba98d
Author: 王开朗 <wangkailang@users.noreply.github.com>
Date:   Thu Nov 10 14:50:57 2022 +0800

    feat: wallet enterprise (#1812)

    * feat: wallet enterprise


    * fix: move back account modal file



* fix: delete embed link error

* fix: add deleted embed link error

* fix: add embed link error

* fix: delete embed link response swagger define

* refactor: save tiny embed link info in database

* fix: check whether the view id exists

* refactor: check whether the view id exists

* feat: add node embed link count limit

* fix: embed link url

* fix: dataPack interface validation error

* fix: error message not completely

* fix: remove hard code

* feat: remove class-validator context param

* fix: api space validate

* test: fix test fail

+++ packages/room-server/src/fusion/fusion.api.controller.ts
@@ -1,4 +1,5 @@
 import { ApiTipConstant, ICollaCommandOptions } from '@apitable/core';
+import { RedisService } from '@apitable/nestjs-redis';
 import {
   Body,
   CacheTTL,
@@ -17,14 +18,14 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiBody, ApiConsumes, ApiInternalServerErrorResponse, ApiOperation, ApiProduces, ApiTags } from '@nestjs/swagger';
-import { RedisService } from '@apitable/nestjs-redis';
 import { AttachmentDto } from 'database/dtos/attachment.dto';
 import { InternalCreateDatasheetVo } from 'database/interfaces';
 import { AttachmentUploadRo } from 'database/ros/attachment.upload.ro';
 import { AttachmentService } from 'database/services/attachment/attachment.service';
 import { FusionApiService } from 'fusion/services/fusion.api.service';
+import { RecordDeleteVo } from 'fusion/vos/record.delete.vo';
 import { I18nService } from 'nestjs-i18n';
-import { DATASHEET_HTTP_DECORATE, NodePermissions, SwaggerConstants, USER_HTTP_DECORATE } from 'shared/common';
+import { API_MAX_MODIFY_RECORD_COUNTS, DATASHEET_HTTP_DECORATE, NodePermissions, SwaggerConstants, USER_HTTP_DECORATE } from 'shared/common';
 import { NodePermissionEnum } from 'shared/enums/node.permission.enum';
 import { ApiException } from 'shared/exception';
 import { RedisLock } from 'shared/helpers/redis.lock';
@@ -63,7 +64,6 @@ import { AssetView, AttachmentVo } from './vos/attachment.vo';
 import { DatasheetCreateDto, DatasheetCreateVo, FieldCreateVo } from './vos/datasheet.create.vo';
 import { FieldDeleteVo } from './vos/field.delete.vo';
 import { FieldListVo } from './vos/field.list.vo';
-import { RecordDeleteVo } from './vos/record.delete.vo';
 import { RecordListVo } from './vos/record.list.vo';
 import { RecordPageVo } from './vos/record.page.vo';
 import { ViewListVo } from './vos/view.list.vo';
@@ -291,6 +291,9 @@ export class FusionApiController {
   @ApiProduces('application/json')
   @UseGuards(ApiDatasheetGuard)
   public async deleteRecord(@Param() param: RecordParamRo, @Query() query: RecordDeleteRo): Promise<RecordDeleteVo> {
+    if (query.recordIds.length > API_MAX_MODIFY_RECORD_COUNTS) {
+      throw ApiException.tipError(ApiTipConstant.api_params_records_max_count_error, { count: API_MAX_MODIFY_RECORD_COUNTS });
+    }
     const result = await this.fusionApiService.deleteRecord(param.datasheetId, query.recordIds);
     if (result) {
       return ApiResponse.success(undefined);
